move auth sync out of Account Provider; remove faq todo
